Use int32 for index and add functions to get the control words

+++ sentencepiece/token.go
@@ -2,12 +2,12 @@ package sentencepiece
 
 // Token holds a unit of a tokenized word
 type Token struct {
-	ID   int64
+	ID   int32
 	Text string
 }
 
 type tokenOffset struct {
-	id    int64
+	id    int32
 	text  string
 	start int
 	end   int
